# Optimization for use_ws_distance=true