--- a/src/postw90/berry.F90
+++ b/src/postw90/berry.F90
@@ -1205,7 +1205,7 @@ contains
     use w90_constants, only: dp, cmplx_0, cmplx_i
     use w90_utility, only: utility_re_tr_prod, utility_im_tr_prod
     use w90_parameters, only: num_wann, nfermi
-    use w90_postw90_common, only: pw90common_fourier_R_to_k_vec, pw90common_fourier_R_to_k
+    use w90_postw90_common, only: pw90common_fourier_R_to_k_vec, pw90common_fourier_R_to_k_new
     use w90_wan_ham, only: wham_get_eig_UU_HH_JJlist, wham_get_occ_mat_list
     use w90_get_oper, only: AA_R, BB_R, CC_R
     use w90_utility, only: utility_zgemm_new
@@ -1313,7 +1313,7 @@ contains
       call pw90common_fourier_R_to_k_vec(kpt, BB_R, OO_true=BB)
       do j = 1, 3
         do i = 1, j
-          call pw90common_fourier_R_to_k(kpt, CC_R(:, :, :, i, j), CC(:, :, i, j), 0)
+          call pw90common_fourier_R_to_k_new(kpt, CC_R(:, :, :, i, j), OO=CC(:, :, i, j))
           CC(:, :, j, i) = conjg(transpose(CC(:, :, i, j)))
         end do
       end do