--- a/src/postw90/geninterp.F90
+++ b/src/postw90/geninterp.F90
@@ -29,7 +29,7 @@ module w90_geninterp
   use w90_get_oper, only: get_HH_R, HH_R
   use w90_comms
   use w90_utility, only: utility_diagonalize
-  use w90_postw90_common, only: pw90common_fourier_R_to_k
+  use w90_postw90_common, only: pw90common_fourier_R_to_k_new
   use w90_wan_ham, only: wham_get_eig_deleig
   use w90_io, only: io_date
   implicit none
@@ -233,7 +233,7 @@ contains
       if (geninterp_alsofirstder) then
         call wham_get_eig_deleig(kpt, localeig(:, i), localdeleig(:, :, i), HH, delHH, UU)
       else
-        call pw90common_fourier_R_to_k(kpt, HH_R, HH, 0)
+        call pw90common_fourier_R_to_k_new(kpt, HH_R, OO=HH)
         call utility_diagonalize(HH, num_wann, localeig(:, i), UU)
       end if
     end do