--- a/test-suite/tests/testpostw90_fe_dos_spin/Fe.mmn.bz2
+++ b/test-suite/tests/testpostw90_fe_dos_spin/Fe.mmn.bz2
@@ -1,0 +1,1 @@
+../../checkpoints/fe_postw90/Fe.mmn.bz2