Dear all,

This PR implements optimization of the use of Wigner-Seitz minimal distance replica selection (`use_ws_distance= true`) in the R to k Fourier transformation.

The optimization has been discussed by Stepan Tsirkin (#296, point 5) and is implemented in WannierBerri also by Stepan.
The WannierBerri website has detailed information on this optimization:
http://wannier-berri.org/methods.html#minimal-distance-replica-selection-method

Briefly, the real-space matrix elements are divided by the degeneracy factors `ndegen` and `wdist_ndeg` only once, in `get_oper` by the routine `operator_wigner_setup`. The `fourier_R_to_k` routines use a single set of R-point data: `nrpts_pw90`, `irvec_pw90`, and `crvec_pw90`, which is common, irrespective of `use_ws_distance`.
Hence, the optimized code is also much simpler than the previous one: the `fourier_R_to_k` routines do _not_ need two branches for `use_ws_distance=true` and `use_ws_distance=false`.

I tested the performance by running the calculation of testpostw90_fe_ahc, with berry_kmesh increased to 50*50*50. I used a single core. "Original" refers to the develop branch, commit 2f4aed6.
* `use_ws_distance=false`, original:  42.806 sec
* `use_ws_distance=true`, original: 136.720 sec
* `use_ws_distance=false`, optimized:  43.496 sec
* `use_ws_distance=true`, optimized:  43.237 sec

With the optimized code, using `use_ws_distance = true` requires no extra computational cost.

Also, I made the following additional modifications:
* Tighten tolerance of the shift current test. The range of the benchmark data in `testpostw90_gaas_sc_xyz` is 0 to 2E-5, but the previous absolute tolerance was 1E-4. I tightened the tolerance to 1E-6.
* Add four tests: `testpostw90_gaas_sc_xyz_scphase2`, `testpostw90_gaas_sc_xyz_scphase2_ws`, `testpostw90_gaas_sc_xyz_wsdistance`, `testpostw90_fe_kpathmorbcurv_ws`. Now, all the fourier_R_to_k routines are covered by tests.
* Add clocks (activated when `timing_level>2`) to some fourier_R_to_k routines
* Move call to `ws_translate_dist` from `pw90common_fourier_R_to_k` to `pw90common_wanint_setup` (This is also done in #275: Technical details, point 2)
* Remove source code of `pw90common_fourier_R_to_k` (the calls are removed in #338)

This PR is written assuming that #338 is merged.

Best regards,
Jae-Mo