@jaemolihm 2020-07-17
Dear Stepan,

No, the writing of "_*.dat" files is not modified by this pull request.
 I only modified the source codes of `postw90.x` (those inside the `src/postw90/` directory), while the _*.dat files are written by `wannier90.x`. 

Jae-Mo