@jaemolihm 2020-07-17
I also should note that the optimization may not work with `effective_model = true and use_ws_distance = true`. The reason is the setup for ws_distance is called in `pw90common_wanint_setup`, before the R vectors are read (which is done in `get_HH_R`).
It seems that  `effective_model = true` is intended not to work with `use_ws_distance = true`. If so, it would be better to have a check in parameters.F90?

I believe `effective_model = true and use_ws_distance = false` should work fine, but I have not tested it, as I could not find a working example..